// ra_automation/data/enums/AdminDashboardBucket.ts
export enum AdminDashboardBucket {
    ROLES = 'ROLE',
    USERS = 'USER',
    DOCTORS = 'DOCTOR', 
    SALES_REP = 'SR', 
    TECHNICIANS = 'TECH', 
    FACILITY = 'FA', 
    FACILITY_GROUP = 'FG', 
    BRANCH_OFFICE = 'BO', 
    INSURANCE_PROVIDER = 'IP', 
    TYPES = 'MT', 
    STATUS = 'MS', 
    SHIPPER = 'MSH', 
    DOCUMENT_TYPE = 'DT', 
    USER_ACTIVITY = 'UACT', 
    NOTE_REASON_TYPE = 'NRT', 
    FEE_SCHEDULE = 'FS', 
    ICD_10 = 'ICD', 
    HCPCS_CODE = 'HCPCS', 
    FAX_EMAIL_CONFIG = 'FEC', 
    WAREHOUSES = 'WH', 
    ORDER_TYPES = 'OT', 
    PDF_REPORT_TEMPLATE = 'PDFRT', 
    SPECIALTY = 'SPC', 
    RESUPPLY_CAMPAIGNS = 'RSL', 
    CAMPAIGN_TYPE = 'CMGN', 
    STANDARD_ACTIVITY = 'SA', 
    MEDICARE_FEE_SCHEDULE = 'MFS',
}