// ra_automation/data/enums/HomepageDashboardBucket.ts
export enum HomepageDashboardBucket {
    INCOMING_FAX = 'INCOMING FAX', 
    NEW_ORDERS = 'NEW ORDERS', 
    PENDING = 'PENDING', 
    CONTACT_PATIENT = 'CONTACT PATIENT', 
    PENDING_DELIVERY = 'PENDING DELIVERY', 
    READY_CONFIRMATION = 'READY CONFIRMATION', 
    COMPLIANCE = 'COMPLIANCE', 
    RESUPPLIES = 'RESUPPLIES', 
    TELEMONITORING = 'TELEMONITORING', 
    SLEEP_MANAGEMENT = 'SLEEP MANAGEMENT'
}